Refactor remaining components to external files
